+++ lib/src/features/dashboard/domain/model/summary.dart
@@ -7,8 +7,8 @@ part 'summary.g.dart';
 class TransactionSummary with _$TransactionSummary {
   const factory TransactionSummary({
     required int month,
-    @Default(0) double income,
-    @Default(0) double expense,
+    @Default(0) int income,
+    @Default(0) int expense,
   }) = _TransactionSummary;
 
   factory TransactionSummary.fromJson(Map<String, dynamic> json) => _$TransactionSummaryFromJson(json);
